// src/app/models/propertyListItem.ts
export interface PropertyListItem {
    id: string;
    url: string;
    title: string;
    price: number;
    pricePerMeter: number;
    floor: number;
    market: string;
    buildingType: string;
    area: number;
    private: boolean;
    webName: number;
    city: string;
    district: string;
    firstSeen: string;
    lastSeen: string;
    snapshotCount: number;
    firstPrice: number;
    priceChange: number;
}